fixed import issues and other improvements
