+++ backend/src/btree/mod.rs
@@ -1,5 +1,3 @@
-mod tests;
-
 use std::sync::Arc;
 use std::sync::RwLock;
