+++ backend/src/command/mod.rs
@@ -1,3 +1,3 @@
-mod command;
+pub mod command;
 
 pub use command::Command;
