+++ backend/src/storage/disk_manager.rs
