+++ database-client/Cargo.toml
@@ -4,6 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-database = { path = "../database" }
+database = { path = "../backend" }
 rustyline = "14.0.0"
 rustyline-derive = "0.10.0"
