--- database/src/lib.rs
@@ -1,9 +1,0 @@
-pub mod btree;
-pub mod command;
-pub mod database_handler;
-pub mod protocol;
-pub mod server;
-pub mod storage;
-
-#[cfg(test)]
-pub mod tests;
